int -> np.int64

+++ util/wan2respack/qe2respack.py
@@ -219,7 +219,7 @@ def symmetry(root, oldxml=False):
         child = root.find('SYMMETRIES')
         n_sym = int(child.find('NUMBER_OF_SYMMETRIES').text)
         ftau = np.zeros((3, n_sym))
-        mat_sym = np.zeros((3, 3, n_sym), np.int)
+        mat_sym = np.zeros((3, 3, n_sym), np.int64)
         for i in range(n_sym):
             sym = child.find('SYMM.{0}'.format(i + 1))
             rot = sym.find('ROTATION').text.strip().split('\n')
@@ -231,7 +231,7 @@ def symmetry(root, oldxml=False):
         child = root.find('output').find('symmetries')
         n_sym = int(child.find('nsym').text)
         ftau = np.zeros((3, n_sym))
-        mat_sym = np.zeros((3, 3, n_sym), int)
+        mat_sym = np.zeros((3, 3, n_sym), np.int64)
         for i, sym in enumerate(child.iter('symmetry')):
             ftau[:, i] = [float(x) for x in sym.find('fractional_translation').text.split()]
             rot = sym.find('rotation').text.strip().split('\n')
